requirements.txt
web3
python-dotenv
fastapi
uvicorn
py-solc-x
